+++ ServiceLibrary/Product.cs
@@ -16,7 +16,7 @@ namespace ServiceLibrary
     {
         public int Id { get; set; }
         public string Name { get; set; }
-        public int Price { get; set; }
-        public double Stock { get; set; }
+        public double Price { get; set; }
+        public int Stock { get; set; }
     }
 }
